app/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy>=2.0
pyodbc
python-dotenv
pydantic
pydantic-settings
python-jose[cryptography]
passlib[bcrypt]
python-multipart